changes added for org details v21.2

+++ src/main/java/com/winwin/winwin/service/OrganizationNoteService.java
@@ -57,13 +57,9 @@ public class OrganizationNoteService implements IOrganizationNoteService {
 				note.setUpdatedBy(user.getEmail());
 				note = organizationNoteRepository.saveAndFlush(note);
 
-				if (null != note.getOrganizationId()) {
-					OrganizationHistory orgHistory = new OrganizationHistory();
-					orgHistory.setOrganizationId(note.getId());
-					orgHistory.setUpdatedAt(sdf.parse(formattedDte));
-					orgHistory.setUpdatedBy(user.getUserDisplayName());
-					orgHistory.setActionPerformed(OrganizationConstants.CREATE_NOTE);
-					orgHistory = orgHistoryRepository.saveAndFlush(orgHistory);
+				if (null != note && null != note.getOrganizationId()) {
+					createOrgHistory(user, note.getOrganizationId(), sdf, formattedDte, OrganizationConstants.CREATE,
+							OrganizationConstants.NOTE, note.getId());
 				}
 			}
 		} catch (Exception e) {
@@ -76,24 +72,19 @@ public class OrganizationNoteService implements IOrganizationNoteService {
 	@Override
 	public void removeOrganizationNote(Long noteId, Long orgId) {
 		UserPayload user = getUserDetails();
+		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+		String formattedDte = sdf.format(new Date(System.currentTimeMillis()));
 
-		if (null != user) {
-			organizationNoteRepository.deleteById(noteId);
-			if (null != orgId) {
-				SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-				String formattedDte = sdf.format(new Date(System.currentTimeMillis()));
-				OrganizationHistory orgHistory = new OrganizationHistory();
-				orgHistory.setOrganizationId(orgId);
-				try {
-					orgHistory.setUpdatedAt(sdf.parse(formattedDte));
-					orgHistory.setUpdatedBy(user.getUserDisplayName());
-					orgHistory.setActionPerformed(OrganizationConstants.DELETE_NOTE);
-					orgHistory = orgHistoryRepository.saveAndFlush(orgHistory);
-				} catch (ParseException e) {
-					LOGGER.error(customMessageSource.getMessage("org.note.error.deleted"), e);
-				}
+		if (null != orgId && null != user) {
+			try {
+				organizationNoteRepository.deleteById(noteId);
 
+				createOrgHistory(user, orgId, sdf, formattedDte, OrganizationConstants.DELETE,
+						OrganizationConstants.NOTE, noteId);
+			} catch (Exception e) {
+				LOGGER.error(customMessageSource.getMessage("org.note.error.deleted"), e);
 			}
+
 		}
 
 	}
@@ -115,6 +106,28 @@ public class OrganizationNoteService implements IOrganizationNoteService {
 
 		}
 		return user;
+	}
+
+	/**
+	 * @param user
+	 * @param orgId
+	 * @param sdf
+	 * @param formattedDte
+	 * @param actionPerformed
+	 * @param entity
+	 * @param entityId
+	 * @throws ParseException
+	 */
+	private void createOrgHistory(UserPayload user, Long orgId, SimpleDateFormat sdf, String formattedDte,
+			String actionPerformed, String entity, Long entityId) throws ParseException {
+		OrganizationHistory orgHistory = new OrganizationHistory();
+		orgHistory.setOrganizationId(orgId);
+		orgHistory.setEntityId(entityId);
+		orgHistory.setEntity(entity);
+		orgHistory.setUpdatedAt(sdf.parse(formattedDte));
+		orgHistory.setUpdatedBy(user.getUserDisplayName());
+		orgHistory.setActionPerformed(actionPerformed);
+		orgHistory = orgHistoryRepository.saveAndFlush(orgHistory);
 	}
 
 }
